This had bugs

+++ include/GeneralColorHist.h
@@ -73,7 +73,7 @@ public:
      *         - color: color index with maximum support
      *         - depth: pattern depth where extension is most promising
      */
-    std::pair<int32_t, int32_t> get_color_to_add(uint32_t threshold=0) ;
+    std::pair<int32_t, int32_t> get_color_to_add(double threshold=0) ;
 
     /**
      * @brief Increase number of trees that stay valid if color is added as neighbour at depth.
